requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgi-redis==0.14.0
asgiref==0.14.0
autobahn==0.16.0
billiard==3.3.0.23
celery==3.1.23
channels==0.17.2
daphne==0.14.3
Django==1.10
djangorestframework==3.4.4
kombu==3.0.35
msgpack-python==0.4.8
py==1.4.31
pytest==3.0.0
pytz==2016.6.1
redis==2.10.5
six==1.10.0
Twisted==16.2.0
txaio==2.5.1
zope.interface==4.2.0